requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.4
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
generativeai==0.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
huggingface-hub==0.26.5
idna==3.10
ipython==8.30.0
jedi==0.19.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
youtube-transcript-api==0.6.3
